+++ src/app/core/auth/auth.service.ts
@@ -24,9 +24,6 @@ export class AuthService {
   public user$ = new BehaviorSubject<User>(null);
   // Userdata, streams with every change
   public userAuthenticated$: Observable<UserData>;
-  // TODO: userForInterceptor$ löschen, wenn userID im Token ist
-  // (wird nur noch im Interceptor verwendet, um die Anfrage auf die UserId zuzulassen)
-  public userForInterceptor$ = new BehaviorSubject<User>(null);
   private tokenExpirationTimer: any;
 
   lnkAfterLogout = this.staticConfig.getRoutingInfo().lnkAfterLogout;
@@ -50,10 +47,10 @@ export class AuthService {
   login(email: string, password: string): Observable<User> {
     return this.apiService.loginUser(email, password).pipe(
       map((serverResponse: AuthResponseData) => {
-        console.table(serverResponse);
+        // console.table(serverResponse);
         const expirationDate = new Date(new Date().getTime() + +serverResponse.expires_in * 1000);
         const decoded = this.getDecodedToken(serverResponse.access_token);
-        console.log('Decoded Token: ', decoded);
+        //console.log('Decoded Token: ', decoded);
         const user = new User(
           decoded.user_id,
           email,
@@ -62,7 +59,6 @@ export class AuthService {
           expirationDate
         );
         this.user$.next(user);
-        this.userForInterceptor$.next(user); //DELETE
         this.cookieDataService.setLocalStorageItem('userData', JSON.stringify(user));
         const expirationDuration =
           new Date(user.tokenExpirationDate).getTime() - new Date().getTime();
@@ -74,7 +70,6 @@ export class AuthService {
 
   logout() {
     this.user$.next(null);
-    this.userForInterceptor$.next(null);
     this.cookieDataService.deleteLocalStorageItem('userData');
     if (this.tokenExpirationTimer) {
       clearTimeout(this.tokenExpirationTimer);
@@ -101,11 +96,10 @@ export class AuthService {
 
     if (!userData || !userData._token) {
       this.user$.next(null);
-      this.userForInterceptor$.next(null);
       return;
     }
 
-    console.log('UserData: ', userData);
+    //console.log('UserData: ', userData);
     //console.log('LocalStorage' + JSON.stringify(localStorage.getItem('userData')));
 
     const user = new User(
@@ -117,7 +111,6 @@ export class AuthService {
     );
 
     this.user$.next(user);
-    this.userForInterceptor$.next(user);
   }
 
   private getDecodedToken(token: string): AuthTokenStructure {
@@ -129,35 +122,35 @@ export class AuthService {
   }
 
   // TO_DELETE
-  login_alt(email: string, password: string): Observable<any> {
-    return this.apiService.loginUser(email, password).pipe(
-      concatMap((serverResponse: AuthResponseData) => {
-        console.table(serverResponse);
-        const expirationDate = new Date(new Date().getTime() + +serverResponse.expires_in * 1000);
-        const user = new User(0, email, undefined, serverResponse.access_token, expirationDate);
+  // login_alt(email: string, password: string): Observable<any> {
+  //   return this.apiService.loginUser(email, password).pipe(
+  //     concatMap((serverResponse: AuthResponseData) => {
+  //       console.table(serverResponse);
+  //       const expirationDate = new Date(new Date().getTime() + +serverResponse.expires_in * 1000);
+  //       const user = new User(0, email, undefined, serverResponse.access_token, expirationDate);
 
-        // Sonst kann das Token über den Interceptor nicht mitgesendet werden
-        this.userForInterceptor$.next(user); // Todo: Delete
+  //       // Sonst kann das Token über den Interceptor nicht mitgesendet werden
+  //        this.userForInterceptor$.next(user); // Todo: Delete
 
-        // Todo: Die UserID und Namen im Token mitsenden,
-        // dann muss man hier keine weiteren Abfragen machen
-        return this.apiService.getUserByEmail(email).pipe(
-          tap((userDataArr) => {
-            //console.log('UserData from Server: ', userDataArr);
-            let userData = userDataArr[0];
-            if (userData) {
-              user.id = userData.id;
-              user.name = userData.name;
-              this.user$.next(user);
-              this.userForInterceptor$.next(user);
-              this.cookieDataService.setLocalStorageItem('userData', JSON.stringify(user));
-              const expirationDuration =
-                new Date(user.tokenExpirationDate).getTime() - new Date().getTime();
-              this.autoLogout(expirationDuration);
-            }
-          })
-        );
-      })
-    );
-  }
+  //       // Todo: Die UserID und Namen im Token mitsenden,
+  //       // dann muss man hier keine weiteren Abfragen machen
+  //       return this.apiService.getUserByEmail(email).pipe(
+  //         tap((userDataArr) => {
+  //           //console.log('UserData from Server: ', userDataArr);
+  //           let userData = userDataArr[0];
+  //           if (userData) {
+  //             user.id = userData.id;
+  //             user.name = userData.name;
+  //             this.user$.next(user);
+  //              this.userForInterceptor$.next(user); //Delete
+  //             this.cookieDataService.setLocalStorageItem('userData', JSON.stringify(user));
+  //             const expirationDuration =
+  //               new Date(user.tokenExpirationDate).getTime() - new Date().getTime();
+  //             this.autoLogout(expirationDuration);
+  //           }
+  //         })
+  //       );
+  //     })
+  //   );
+  // }
 }